# api/src/main/resources/assets/openapi.yml
...
openapi: 3.0.2
info:
  title: Marquez
  version: 0.51.2-SNAPSHOT
  description: Marquez is an open source **metadata service** for the **collection**, **aggregation**,
    and **visualization** of a data ecosystem's metadata.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
    description: Current API server
paths:
  /api/v1/namespaces/{namespace}:
    parameters:
      - $ref: '#/components/parameters/namespace'
    put:
      operationId: putNamespace
      summary: Create a namespace
      description: Creates a new namespace object. A namespace enables the contextual grouping of related
        jobs and datasets. Namespaces must contain only letters (`a-z`, `A-Z`), numbers (`0-9`),
        underscores (`_`), dashes (`-`), colons (`:`), slashes (`/`), or dots (`.`).
        A namespace is case-insensitive with a maximum length of `1024` characters.
        Note jobs and datasets will be unique within a namespace, but not across namespaces.
      tags:
        - Namespaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedNamespace'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
    get:
      operationId: getNamespace
      summary: Retrieve a namespace
      description: Returns a namespace.
      tags:
        - Namespaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
    delete:
      operationId: deleteNamespace
      summary: Deletes a namespace
      description: Soft deletes a namespace, and every job and dataset inside. On next event containing this namespace, the namespace will be undeleted.
      tags:
        - Namespaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
  /api/v1/namespaces:
    get:
      operationId: getNamespaces
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all namespaces
      description: Returns a list of namespaces.
      tags:
        - Namespaces
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceList'
  /api/v1/events/lineage:
    get:
      operationId: getLineageEvents
      parameters:
        - $ref: '#/components/parameters/sortDirection'
        - $ref: '#/components/parameters/before'
        - $ref: '#/components/parameters/after'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all received OpenLineage events.
      description: Returns a list of OpenLineage events, sorted in direction of passed sort parameter. By default it is desc.
      tags:
        - Events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageEventList'
  /api/v1/sources/{source}:
    parameters:
      - $ref: '#/components/parameters/source'
    put:
      operationId: putSource
      summary: Create a source
      description: Creates a new source object. A source is the physical location of a dataset such as
        a table in PostgreSQL, or topic in Kafka. A source enables the grouping of physical datasets
        to their physical source.
      deprecated: true
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedSource'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
    get:
      operationId: getSource
      summary: Retrieve a source
      description: Returns a source.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /api/v1/sources:
    get:
      operationId: getSources
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all sources
      description: Returns a list of sources.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
...
  /api/v1/namespaces/{namespace}/datasets:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
    get:
      operationId: getDatasets
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all datasets
      description: Returns a list of datasets.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
...
  /api/v1/namespaces/{namespace}/jobs:
    parameters:
      - $ref: '#/components/parameters/namespace'
    get:
      operationId: getJobs
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all jobs
      description: Returns a list of jobs.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
...
  /api/v1/lineage:
    post:
      operationId: recordLineage
      summary: Record a single lineage event
      description: Receive, process, and store lineage metadata using the [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.json) standard.
      tags:
        - Lineage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineageEvent'
      responses:
        '200':
          description: OK
    get:
      operationId: getLineage
      parameters:
        - $ref: '#/components/parameters/nodeId'
        - $ref: '#/components/parameters/depth'
        - $ref: '#/components/parameters/aggregateToParentRun'
        - $ref: '#/components/parameters/includeFacets'
      tags:
        - Lineage
      summary: Get a lineage graph
      description: Returns a lineage graph for the specified node. Optionally aggregate child runs to parent and filter facets to prevent large response sizes.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
  /api/v2/lineage:
    get:
      operationId: getLineageV2
      parameters:
        - $ref: '#/components/parameters/nodeId'
        - $ref: '#/components/parameters/depth'
        - $ref: '#/components/parameters/aggregateToParentRun'
      tags:
        - Lineage
      summary: Get a denormalized lineage graph
      description: Returns a lineage graph for the specified node using the V2 denormalized read path.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
...
  /api/v1/tags/{tag}:
    parameters:
      - $ref: '#/components/parameters/tag'
    put:
      operationId: putTag
      summary: Create a tag
      description: Creates a new tag object.
      tags:
        - Tags
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedTag'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
  /api/v1/tags:
    get:
      operationId: getTags
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all tags
      description: Returns a list of tags.
      tags:
        - Tags
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
...
  /api/v3/namespaces:
    get:
      operationId: getNamespacesV3
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all namespaces (V3 Graph)
      description: Returns a list of namespaces using the V3 Graph backend.
      tags:
        - Namespaces V3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceList'
...
  /api/v3/lineage:
    get:
      operationId: getLineageV3
      parameters:
        - $ref: '#/components/parameters/nodeId'
        - $ref: '#/components/parameters/depth'
      tags:
        - Lineage V3
      summary: Get a lineage graph (V3 Graph)
      description: Returns a lineage graph for the specified node using the V3 Apache AGE backend.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
...
components:
  parameters:
...
    namespace:
      name: namespace
      in: path
      description: The name of the namespace.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-namespace
    source:
      name: source
      in: path
      description: The name of the source.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-source
    dataset:
      name: dataset
      in: path
      description: The name of the dataset.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-dataset
    tag:
      name: tag
      in: path
      description: The name of the tag.
      required: true
      schema:
        type: string
        example: SENSITIVE
...
    limit:
      name: limit
      in: query
      description: The number of results to return from offset.
      required: false
      schema:
        type: integer
        example: 25
        default: 100
    offset:
      name: offset
      in: query
      description: The initial position from which to return results.
      required: false
      schema:
        type: integer
        default: 0
    nodeId:
      name: nodeId
      in: query
      schema:
        type: string
        example: dataset:food_delivery:public.delivery_7_days
      description: The ID of the node. A node can either be a dataset node, a dataset field node or a job node.
        The format of nodeId for dataset is `dataset:<namespace_of_dataset>:<name_of_the_dataset>`, for dataset field
        is `datasetField:<namespace_of_dataset>:<name_of_the_dataset>:<name_of_field>`, and for job is
        `job:<namespace_of_the_job>:<name_of_the_job>`.
      required: true
    depth:
      name: depth
      in: query
      schema:
        type: integer
        default: 20
      description: Depth of lineage graph to create.
      required: false
...
    aggregateToParentRun:
      name: aggregateToParentRun
      in: query
      schema:
        type: boolean
        default: false
      description: When true, aggregates all child runs to their parent run UUID, grouping facets and lineage data together. Useful for jobs with dynamic child tasks.
      required: false
    includeFacets:
      name: includeFacets
      in: query
      schema:
        type: array
        items:
          type: string
        example: ["spark", "processing_engine"]
      description: Filters run facets to include only the specified facet names. Reduces response size and prevents PostgreSQL 256MB JSONB limit issues with large lineage queries. Facets are filtered at SQL level for optimal performance.
      required: false
      style: form
      explode: true
...
    sortDirection:
      name: sortDirection
      in: query
      schema:
        type: string
        example: name
        description: Sorts the results of your query by indicated direction `asc` or `desc`.
      required: false
    before:
      name: before
      in: query
      description: Returns events before passed date.
      required: false
      schema:
        type: string
        format: date-time
        example: 2022-09-15T07:47:19Z
    after:
      name: after
      in: query
      description: Returns events after passed date.
      required: false
      schema:
        type: string
        format: date-time
        example: 2022-09-15T07:47:19Z
...
  schemas:
    CreatedNamespace:
      type: object
      properties:
        ownerName:
          description: The owner of the namespace.
          type: string
        description:
          description: The description of the namespace.
          type: string
      required:
        - ownerName
      example:
        ownerName: me
        description: My first namespace!
    Namespace:
      type: object
      properties:
        name:
          description: The name of the namespace.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the namespace was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the namespace was updated.
          type: string
          format: date-time
        ownerName:
          description: The owner of the namespace.
          type: string
        description:
          description: The description of the namespace.
          type: string
      example:
        name: my-namespace
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        ownerName: me
        description: My first namespace!
    NamespaceList:
      type: object
      properties:
        namespaces:
          type: array
          items:
            $ref: '#/components/schemas/Namespace'
    LineageEventList:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/LineageEvent'
        totalCount:
          type: number
          description: The total number of events returned matching our conditions.
    CreatedSource:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the source.
          type: string
      required:
        - type
        - connectionUrl
      example:
        type: POSTGRESQL
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    Source:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        name:
          description: The name of the source.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was updated.
          type: string
          format: date-time
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the dataset.
          type: string
      example:
        type: POSTGRESQL
        name: my-source
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    SourceList:
      type: object
      properties:
        sources:
          items:
            $ref: '#/components/schemas/Source'
...
    DatasetId:
      type: object
      description: The ID of the dataset.
      properties:
        namespace:
          type: string
          description: The namespace of the dataset.
        name:
          type: string
          description: The name of the dataset.
    Dataset:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/DatasetId'
        type:
          description: The type of the dataset.
          type: string
        name:
          description: The **logical** name of the dataset.
          type: string
        physicalName:
          description: The **physical** name of the dataset.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was updated.
          type: string
          format: date-time
        namespace:
          type: string
          description: The namespace of the dataset.
        sourceName:
          description: The name of the source associated with the dataset.
          type: string
        fields:
          description: The fields of the dataset.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The data type of the field.
                type: string
              tags:
                description: List of tags.
                type: array
                items:
                  type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: array
          items:
            type: string
        lastModifiedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was last modified by a successful run.
          type: string
          format: date-time
        lastLifecycleState:
          description: The last lifecycle state of the dataset.
          type: string
        description:
          description: The description of the dataset.
          type: string
        facets:
          $ref: '#/components/schemas/DatasetFacets'
        currentVersion:
          description: The current version of the dataset.
          type: string
          format: uuid
        deleted:
          description: The deleted state of the dataset.
          type: boolean
      example:
        id:
          namespace: my-namespace
          name: my-dataset
        type: DB_TABLE
        name: my-dataset
        physicalName: public.mytable
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        sourceName: my-source
        fields: [
          {name': a, type: INTEGER, tags: []},
          {name': b, type: TIMESTAMP, tags: []},
          {name': c, type: INTEGER, tags: []},
          {name': d, type: INTEGER, tags: []}
        ]
        tags: []
        lastModifiedAt: null
        description: "My first dataset!"
        facets: {}
        currentVersion: "b1d626a2-6d3a-475e-9ecf-943176d4a8c6"
    DatasetList:
      type: object
      properties:
        datasets:
          items:
            $ref: '#/components/schemas/Dataset'
        totalCount:
          type: number
          description: The total number of datasets for the given namespace
      required:
        - datasets
...
    Tag:
      type: object
      properties:
        name:
          description: The name of the tag.
          type: string
        description:
          description: The description of the tag.
          type: string
      example:
        name: SENSITIVE
        description: Contains sensitive information
    TagList:
      type: object
      properties:
        tags:
          items:
            $ref: '#/components/schemas/Tag'
...
    JobList:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        totalCount:
          type: number
          description: The total number of jobs for the given namespace
...
    DatasetFacets:
      description: The facets of the dataset. For a full list of standard dataset facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
...
    CustomFacet:
      description: A custom facet enables the extension of _dataset_, _job_, and _run_ metadata. A custom facet **must** also have a schema,
        where a version of the schema is identifiable via a URL. A field within the schema **must** not start with an underscore (`_`).
      allOf:
        - $ref: '#/components/schemas/BaseFacet'
        - type: object
          additionalProperties: true
...
    LineageGraph:
      type: object
      properties:
        graph:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
    GraphNode:
      type: object
      properties:
        id:
          type: string
          description: The ID of the node.
        type:
          $ref: '#/components/schemas/NodeType'
        data:
          type: object
          description: The data of the node.
          discriminator:
            propertyName: type
            mapping:
              'DATASET': '#/components/schemas/Dataset'
              'JOB': '#/components/schemas/Job'
          oneOf:
            - $ref: '#/components/schemas/Dataset'
            - $ref: '#/components/schemas/Job'
        inEdges:
          type: array
          description: The in edges of the node.
          items:
            $ref: '#/components/schemas/GraphEdge'
        outEdges:
          type: array
          description: The out edges of the node.
          items:
            $ref: '#/components/schemas/GraphEdge'
    NodeType:
      type: string
      enum: [ JOB, DATASET ]
      description: The type of the diff node.
    GraphEdge:
      type: object
      properties:
        origin:
          description: The ID of the origin node.
          type: string
        destination:
          description: The ID of the destination node.
          type: string
    CreatedTag:
      type: object
      properties:
        description:
          description: The description of the tag.
          type: string
      example:
        description: My first tag!
...
    LineageEvent:
      example:
        eventType: COMPLETE
        eventTime: 2019-05-09T19:49:24.201361Z
        run: {runId: d46e465b-d358-4d32-83d4-df660ff614dd}
        job: {namespace: my-namespace,name: my-job}
        outputs: [
          {
            namespace: my-namespace,
            name: my-output,
            facets: {
              schema: {
                _producer: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client",
                _schemaURL: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/spec/OpenLineage.json#/definitions/SchemaDatasetFacet",
                fields: [
                  { name: a, type: INTEGER },
                  { name: b, type: TIMESTAMP },
                  { name: c, type: INTEGER },
                  { name: d, type: INTEGER }
                ]
              }
            }
          }
        ]
        producer: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client"
...